Adds activity selection to evaluation dialog
Enables the selection of activities within the evaluation dialog.

Moves the `Evaluation` model to the evaluations module.

Introduces an "Activity" selection dropdown in the evaluation dialog,
allowing users to associate an activity with the evaluation.

Adds functionality to create a new activity directly from the evaluation
dialog if no activities are available.

Refactors the location of the `Evaluation` model and its spec file
to reside within the evaluations module for better organization.