Improves drag and drop group management
Enhances the drag and drop functionality for managing students within groups. It ensures the group list is reversed for better UX and that group updates are properly handled after a student is dropped into a group. Additionally, it adds a group ID factory and uses ViewChildren to improve the connection between drag lists.
as far we store students in group, ww have to show them

This addresses issues with group updates not being triggered correctly after drag and drop.