Orders user list by last name
Updates the user list to be sorted alphabetically by last name, then first name.
This improves the user experience by making it easier to find specific users in the list.
Also, it corrects the display of the user's full name on the user card,
displaying the last name before the first name.