Refactors user activation process
Simplifies and centralizes the user activation process by introducing a dedicated function.

This change removes the direct email sending logic from the `createUserPlus` function and introduces a new `sendUserActivationLink` function to handle the generation and sending of activation links. This improves code modularity and reusability.
The activation link is only sent if the user is new or the email is not yet verified.