Renames `classe` form control to `classKey`
Renames the `classe` form control to `classKey` in the user generalities component to align with the backend data structure. it fixes a bug introduced by the agent who renamed some variables

Fetches and sets the user's classes based on the logged-in user's data.
Includes console logs for debugging user and class data.