Refactors data fetching for improved performance
Improves the performance of data fetching operations by utilizing asynchronous calls with `Promise.all` when retrieving related entities like classes.

This prevents blocking operations and allows for parallel data retrieval, resulting in a smoother and more responsive user experience, especially when dealing with multiple classes. It also adds a cache to the ClassiService to reduce calls to Firestore.