Refactors classe-dialog page with sidebar wip
Replaces tabs with a sidebar navigation for better user experience on the classe-dialog page.
there are some problerms the menu does not open  always

This change aims to improve the navigation and layout of the classe-dialog, making it more intuitive and user-friendly, especially on smaller screens. The previous tab-based navigation has been replaced with a sidebar, offering a clearer and more accessible menu structure. This also includes adding an overlay to close the sidebar and styling improvements.